+++ cli/worker.rs
@@ -110,29 +110,13 @@ impl Worker {
 
     let global_state_ = state.borrow().global_state.clone();
 
-    let inspector = if let Some(inspector_server) =
-      global_state_.inspector_server.as_ref()
-    {
-      use deno_core::v8;
-      let deno_core::Isolate {
-        v8_isolate,
-        global_context,
-        ..
-      } = &mut **isolate;
-      let isolate_handle = v8_isolate.as_mut().unwrap().thread_safe_handle();
-      let mut hs = v8::HandleScope::new(v8_isolate.as_mut().unwrap());
-      let scope = hs.enter();
-      let context = global_context.get(scope).unwrap();
-
-      let inspector_rx = inspector_server.add_inspector(isolate_handle.clone());
-
-      let inspector =
-        crate::inspector::DenoInspector::new(scope, context, inspector_rx);
-
-      Some(inspector)
-    } else {
-      None
-    };
+    let inspector =
+      if let Some(inspector_server) = global_state_.inspector_server.as_ref() {
+        let inspector = inspector_server.add_inspector(&mut *isolate);
+        Some(inspector)
+      } else {
+        None
+      };
 
     isolate.set_js_error_create_fn(move |core_js_error| {
       JSError::create(core_js_error, &global_state_.ts_compiler)
